+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/notify/MetaResponseService.java
@@ -2,7 +2,7 @@ package com.dwarfeng.familyhelper.webapi.stack.service.notify;
 
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.notify.DispMeta;
 import com.dwarfeng.notify.stack.bean.entity.Meta;
-import com.dwarfeng.notify.stack.bean.entity.key.MetaKey;
+import com.dwarfeng.notify.stack.bean.key.MetaKey;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;